% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{cv.delta}
\alias{cv.delta}
\title{Cross-validation function to select regularization parameter delta in the penalized loss criterion.}
\usage{
cv.delta(
  y1,
  X1,
  z.names,
  K = 10,
  est.MSE = c("est.var", "step")[1],
  show.plots
)
}
\arguments{
\item{y1}{(n by 1) matrix corresponding to the response variable.}

\item{X1}{(n by L) matrix of main covariates where n is the sample size and L=m if z is NULL,
and L= m+1 otherwise. Here, m refers to the number of x-covariates.}

\item{z.names}{character denoting the column name of the z-covariate if z is not NULL. Can be NULL.}

\item{K}{scalar denoting the number of folds for cross-validation
when \code{penalty.choice} is "cv.mse" or "cv.penalized.loss". Default is 10.}

\item{est.MSE}{character that indicates how the mean squared error is estimated in the penalized loss
criterion when \code{penalty.choice} is "penalized.loss" or "cv.penalized.loss". Options are
"est.var" which means the MSE is sd(y) * sqrt(n/(n-1)) where n is the sample size, and
"step" which means we use the MSE from forward stepwise regression with AIC as the selection criterion. Default
is "est.var".}

\item{show.plots}{logical indicator. If TRUE and \code{penalty.choice} is "penalized.loss", a plot of the penalized
loss criterion versus steps in the LARS algorithm of Efron et al (2004).}
}
\value{
\itemize{
  \item \strong{predict.out:}{output from LARS algorithm of Efron et al (2004).}
  \item \strong{delta.out:}{the penalization parameter delta obtained from K-fold cross validation.}
}
}
\description{
Cross-validation function to select regularization parameter delta in the penalized loss criterion.
}
\references{
Efron, B., Hastie, T., Johnstone, I. AND Tibshirani, R. (2004). Least angle regression.
Annals of Statistics 32, 407–499.

Garcia, T.P. and M¨uller, S. (2014). Influence of measures of significance-based weights in the weighted Lasso.
Journal of the Indian Society of Agricultural Statistics (Invited paper), 68, 131-144.

Garcia, T.P., Mueller, S., Carroll, R.J., Dunn, T.N., Thomas, A.P., Adams, S.H., Pillai, S.D., and Walzem, R.L.
(2013). Structured variable selection with q-values. Biostatistics, DOI:10.1093/biostatistics/kxt012.
}
