% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{get.weight.fn}
\alias{get.weight.fn}
\title{Weight function examples}
\usage{
get.weight.fn(
  weight_fn_type = c("identity", "sqrt", "inverse_abs", "square")[1]
)
}
\arguments{
\item{weight_fn_type}{character indicating type of function to return.}
}
\value{
a function that can be applied to a scalar.
}
\description{
Returns possible weight functions that can be used in d2wlasso.
}
\examples{
weight_fn <- get.weight.fn("sqrt") # return f(x)=sqrt(x)

}
