% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{parcorr.pvalue}
\alias{parcorr.pvalue}
\title{Partial correlation, p-value and t-statistic associated with the regression between response y and a covariate x
after adjustment for a covariate z.}
\usage{
parcorr.pvalue(
  factor.z,
  x,
  y,
  z,
  delta = NULL,
  method = "pearson",
  alternative = "two.sided",
  ttest.pvalue = FALSE,
  regression.type
)
}
\arguments{
\item{factor.z}{logical. If TRUE, the fixed variable z is a factor variable.}

\item{x}{(n by 1) matrix corresponding to a covariate vector where n is the sample size.}

\item{y}{(n by 1) a matrix corresponding to the response variable. If \code{regression.type} is "cox",
\code{y} contains the observed event times.}

\item{z}{(n by 1) matrix of additional fixed covariate affecting response variable. Can be NULL.}

\item{delta}{(n by 1) a matrix that denotes censoring when \code{regression.type} is "cox" (1 denotes
survival event is observed, 0 denotes the survival event is censored). Can be NULL.}

\item{method}{character indicating the type of correlation to compute. Default if "pearson"}

\item{alternative}{character indicating whether the p-value is computed using one-sided or two-sided
testing. Default is "two-sided".}

\item{ttest.pvalue}{logical indicator. If TRUE, p-value for each covariate is computed from univariate
linear/cox regression of the response on each covariate. If FALSE, the
p-value is computed from correlation coefficients between the response and each covariate.
Default is FALSE.}

\item{regression.type}{a character indicator that is either "linear" for linear regression
or "cox" for Cox proportional hazards regression. Default is "linear".}
}
\value{
\itemize{
  \item \strong{p.value:}{p-value of the coefficient of x in the regression of y on x and z.}
  \item \strong{estimate:}{Partial correlation between x and y after adjustment for z.}
  \item \strong{t.stat:}{t-statistic with testing the significance of x in the regression
  of y on x and z.}
}
}
\description{
Partial correlation, p-value and t-statistic associated with the regression between response y and a covariate x
after adjustment for a covariate z.
}
\examples{
x = matrix(rnorm(100, 0, 1),100,1)
colnames(x) <- paste0("X",1:ncol(x))
z = matrix(rbinom(100, 1, 0.5),100,1)
y = matrix(z[,1] + 2*x[,1] + rnorm(100, 0, 1), 100)
colnames(y) <- "y"
delta <- NULL

output <- parcorr.pvalue(factor.z=TRUE,x,y,z,delta,regression.type="linear")

}
