% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{store.xy}
\alias{store.xy}
\title{Matrix to store results for covariates x and vector response y.}
\usage{
store.xy(XX, yy)
}
\arguments{
\item{XX}{(n by m) matrix of main covariates where m is the number of covariates and n is the sample size.}

\item{yy}{(n by 1) a matrix corresponding to the response variable.}
}
\value{
(m by 1) matrix of zeroes
}
\description{
Forms a matrix that will be used to store regression results of response y on covariates x.
}
\examples{

xx = matrix(rnorm(100*5, 0, 1),100,5)
colnames(xx) <- paste0("X",1:ncol(xx))
yy = matrix(2*xx[,1] - 2*xx[,2] + rnorm(100, 0, 1), 100)
colnames(yy) <- "y"
store.xy(xx,yy)
}
