% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{weighted.lasso}
\alias{weighted.lasso}
\title{Compute weighted lasso variable selection}
\usage{
weighted.lasso(
  weights,
  weight_fn = function(x) {     x },
  yy,
  XX,
  z,
  data.delta,
  z.names,
  show.plots = FALSE,
  penalty.choice,
  est.MSE = c("est.var", "step")[2],
  cv.folds = 10,
  delta = 2
)
}
\arguments{
\item{weights}{(m x 1) matrix that we use to multiply the m-covariates by.}

\item{weight_fn}{A user-defined function to be applied to the weights for the weighted lasso.
Default is an identify function.}

\item{yy}{(n by 1) a matrix corresponding to the response variable. If \code{regression.type} is "cox",
\code{yy} contains the observed event times.}

\item{XX}{(n by K) matrix of main covariates where n is the sample size and K=m if z is NULL,
and K= m+1 otherwise. Here, m refers to the number of x-covariates.}

\item{z}{(n by 1) matrix of additional fixed covariate affecting response variable. This covariate will
always be selected. Can be NULL.}

\item{data.delta}{(n by 1) a matrix that denotes censoring when \code{regression.type} is "cox" (1 denotes
survival event is observed, 0 denotes the survival event is censored). Can be NULL.}

\item{z.names}{character denoting the column name of the z-covariate if z is not NULL. Can be NULL.}

\item{show.plots}{logical indicator. If TRUE and \code{penalty.type} is "penalized.lasso",
a plot of the penalized lasso criterion versus steps in the LARS algorithm of Efron et al (2004).}

\item{penalty.choice}{character that indicates the variable selection criterion. Options are "cv.mse" for
the K-fold cross-validated mean squared prediction error, "penalized.loss" for the penalized loss criterion which
requires specification of the penalization parameter \code{penalized.loss.delta},
"cv.penalized.loss" for the K-fold cross-validated criterion to determine delta in the penalized loss
criterion, and "deviance.criterion" for optimizing the
Cox proportional hazards deviance (only available when \code{regression.type} is "cox".) Defalt is "penalized.loss".}

\item{est.MSE}{character that indicates how the mean squared error is estimated in the penalized loss
criterion when \code{penalty.choice} is "penalized.loss" or "cv.penalized.loss". Options are
"est.var" which means the MSE is sd(y) * sqrt(n/(n-1)) where n is the sample size, and
"step" which means we use the MSE from forward stepwise regression with AIC as the selection criterion. Default
is "est.var".}

\item{cv.folds}{scalar denoting the number of folds for cross-validation
when \code{penalty.choice} is "cv.mse" or "cv.penalized.loss". Default is 10.}

\item{delta}{scalar to indicate the choice of the penalization parameter delta in the
penalized loss criterion when \code{penalty.choice} is "penalized.loss".}
}
\value{
\itemize{
  \item \strong{sig.variables:}{m-dimensional vector where the kth entry is 1 if x_k is selected to be
  in the model, and 0 if not.}
  \item \strong{sign.of.variables:}{m-dimensional vector where the kth entry is 1 if the coefficient in
  front of x_k is positive, and -1 if not.}
  \item \strong{delta.out:}{the penalization parameter delta used in the penalized loss criterion.
  When \code{penalty.choice} is "penalized.loss", \code{delta.out} is the same as \code{delta}.
  When \code{penalty.choice} is "cv.penalized.loss", \code{delta.out} is the resulting delta-value
  obtained from the k-fold cross validation.}
}
}
\description{
Performs variable selection with covariates multiplied by weights that direct which variables
are likely to be associated with the response.
}
\references{
Efron, B., Hastie, T., Johnstone, I. AND Tibshirani, R. (2004). Least angle regression.
Annals of Statistics 32, 407–499.

Garcia, T.P. and M¨uller, S. (2016). Cox regression with exclusion frequency-based weights to
identify neuroimaging markers relevant to Huntington’s disease onset. Annals of Applied Statistics, 10, 2130-2156.

Garcia, T.P. and M¨uller, S. (2014). Influence of measures of significance-based weights in the weighted Lasso.
Journal of the Indian Society of Agricultural Statistics (Invited paper), 68, 131-144.

Garcia, T.P., Mueller, S., Carroll, R.J., Dunn, T.N., Thomas, A.P., Adams, S.H., Pillai, S.D., and Walzem, R.L.
(2013). Structured variable selection with q-values. Biostatistics, DOI:10.1093/biostatistics/kxt012.

Storey, J. D. and Tibshirani, R. (2003). Statistical significance for genomewide studies.
Proceedings of the National Academy of Sciences 100, 9440-9445.
}
