% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2wlasso.R
\name{d2wlasso}
\alias{d2wlasso}
\title{Title}
\usage{
d2wlasso(x, z, y)
}
\arguments{
\item{x}{(n by m1) matrix of main covariates}

\item{z}{(n by m0) matrix of additional fixed covariates affecting response variable}

\item{y}{vector of response variable}
}
\value{
sumx: sum of main covariates

sumz: sum of additional fixed covariates
}
\description{
Title
}
\examples{
x=matrix(rnorm(100*5, 0, 1),100,5)
z=matrix(rnorm(100*1, 0, 1),100,1)
y=2*x[,1] - (2+2*z[,1])*x[,2] + rnorm(100, 0, 1)
d2wlasso(x,z,y)
}
