% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2wlasso.R
\name{d2wlasso}
\alias{d2wlasso}
\title{Implement structured variables selection with q-values}
\usage{
d2wlasso(x, z, y, ttest = FALSE, plots = FALSE, pi0.true = FALSE,
  pi0.val = 0.9)
}
\arguments{
\item{x}{(n by m1) matrix of main covariates where m1 is the number of covariates and n is the sample size}

\item{z}{(n by 1) matrix of additional fixed covariate affecting response variable}

\item{y}{(n by 1) matrix of response variable}
}
\value{
cor.out

parcor.out
}
\description{
Implement structured variables selection with q-values
}
\examples{
x=matrix(rnorm(100*5, 0, 1),100,5)
z <- matrix(rbinom(100, 1, 0.5),100,1)
y=matrix(z[1,] + 2*x[1,] - 2*x[2,] + rnorm(100, 0, 1), 100)
d2wlasso(x,z,y)
}
