% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d2wlasso.R, R/main_functions.R
\docType{package}
\name{d2wlasso}
\alias{d2wlasso}
\title{d2wlasso package}
\usage{
d2wlasso(
  x,
  z,
  y,
  cox.delta = NULL,
  factor.z = TRUE,
  regression.type = c("linear", "cox")[1],
  weight.type = c("one", "corr.estimate", "corr.pvalue", "corr.bh.pvalue", "corr.tstat",
    "corr.qvalue", "parcor.estimate", "parcor.pvalue", "parcor.bh.pvalue", "parcor.tstat",
    "parcor.qvalue", "exfrequency.random.partition.aic",
    "exfrequency.random.partition.bic", "exfrequency.kmeans.partition.aic",
    "exfrequency.kmeans.partition.bic", "exfrequency.kquartiles.partition.aic",
    "exfrequency.kquartiles.partition.bic", "exfrequency.ksorted.partition.aic",
    "exfrequency.ksorted.partition.bic")[1],
  weight_fn = function(x) {     x },
  ttest.pvalue = TRUE,
  q_opt_tuning_method = c("bootstrap", "smoother")[2],
  qval.alpha = 0.15,
  alpha.bh = 0.05,
  robust = TRUE,
  show.plots = FALSE,
  pi0.known = FALSE,
  pi0.val = 0.9,
  penalty.choice = c("cv.mse", "cv.penalized.loss", "penalized.loss",
    "deviance.criterion")[1],
  est.MSE = c("est.var", "step")[1],
  cv.folds = 10,
  mult.cv.folds = 100,
  penalized.loss.delta = 2,
  nboot = 100,
  k.split = 4,
  step.direction = "backward"
)
}
\arguments{
\item{x}{(n by m) matrix of main covariates where m is the number of covariates and n is the sample size.}

\item{z}{(n by 1) matrix of additional fixed covariate affecting response variable. This covariate should
always be selected. Can be NULL.}

\item{y}{(n by 1) a matrix corresponding the response variable. If \code{regression.type} is "cox",
\code{y} contains the observed event times.}

\item{cox.delta}{(n by K) a matrix that denotes censoring when \code{regression.type} is "cox" (1 denotes
survival event is observed, 0 denotes the survival event is censored). Can be NULL.}

\item{factor.z}{logical. If TRUE, the fixed variable z is a factor variable.}

\item{regression.type}{a character indicator that is either "linear" for linear regression
or "cox" for Cox proportional hazards regression. Default is "linear".}

\item{weight.type}{The weights to be used for the weighted lasso. One of "one","t_val","parcor","p_val","bhp_val","adapt","q_cor" or "q_parcor".
"one" gives no weight. "t_val" gives weight of the inverse absolute t-statistics of the regression coefficients. "parcor" gives weight of the inverse absolute partial correlation between the main covariate and the response after accounting for z. "p_val" gives p-value of each predictor's coefficient as weights. "bhp_val" gives Benjamini-Hochberg adjusted p-value of each predictor's coefficient as weights. "adapt" gives adaptive lasso weights, that is, the inverse of the absolute value of regression coefficients. "q_cor" gives weights set to q-values BEFORE taking into account diet. "q_parcor" gives weights set to q-values AFTER taking into account diet.}

\item{weight_fn}{The function applied to the weights for the weighted lasso. One of "identity","sqrt","inverse_abs","square". "identity" is the identity function, "sqrt" is the square root function, "inverse_abs" is the inverse of the absolute value and "square" is the square function. Not used if wt is set to "adapt". Default is "identity".}

\item{ttest.pvalue}{logical. If TRUE, p-value for each covariate is computed from univariate
linear/cox regression of the response on each covariate. If FALSE, the
p-value is computed from correlation coefficients. Default is FALSE.}

\item{q_opt_tuning_method}{indicates the method for choosing optimal tuning parameter
in the q-value computation as proposed in Storey and Tibshirani (2003).
One of "bootstrap" or "smoother". Default is "smoother" (smoothing spline).}

\item{qval.alpha}{indicates cut-off for q-values (thresholding). That is, the covariate with q-value less than this cut-off is included in the model.}

\item{alpha.bh}{indicates cut-off for Benjamini-Hochberg adjusted p-value (thresholding). That is, the covariate with BH-adjusted p-value less than this cut-off is included in the model.}

\item{robust}{indicates whether it is desired to make the estimate more robust for small p-values.}

\item{show.plots}{logical. If TRUE, figures are plotted. Default is FALSE.}

\item{pi0.known}{logical. If TRUE, the estimate of the true proportion of the null hypothesis is set to the value of pi0.val which is given by the user. If FALSE, the estimate of the true proportion of the null hypothesis is computed by bootstrap or smoothing spline. Default is FALSE.}

\item{pi0.val}{A user supplied estimate of the true proportion of the null hypothesis. Used only when pi0.known is TRUE. Default is 0.9.}

\item{penalty.choice}{Among the lasso solution path, the best descriptive model is the one which minimizes the loss function: (residual sum of squares)/(estimator of the model error variance) - (sample size) + delta*(number of predictors in the selected model). If delta = 2, this loss function is Mallows' Cp.}

\item{est.MSE}{indicates the direction of stepwise regression for the Cox regression with exclusion frequency-based weights. One of "both", "forward" or "backward". Default is "backward".}

\item{cv.folds}{indicates the number of folds of the cross-validation for selecting delta.}

\item{mult.cv.folds}{indicates the number of cross-validation runs for selecting delta.}

\item{penalized.loss.delta}{Among the lasso solution path, the best descriptive model is the one which minimizes the loss function: (residual sum of squares)/(estimator of the model error variance) - (sample size) + delta*(number of predictors in the selected model). If delta = 2, this loss function is Mallows' Cp.}

\item{nboot}{indicates the number of bootstrap samples for the Cox regression with exclusion frequency-based weights}

\item{k.split}{indicates the number of partitions for the Cox regression with exclusion frequency-based weights. Default is 4.}

\item{step.direction}{indicates the direction of stepwise regression for the Cox regression with exclusion frequency-based weights. One of "both", "forward" or "backward". Default is "backward".}
}
\value{
\itemize{
   \item \strong{qval:} {q-value as proposed in Storey and Tibshirani (2003)}
   \item \strong{bh.pval:} {Benjamini-Hochberg adjusted p-value as proposed in Benjamini and Hochberg (1995)}
   \item \strong{pval:} {p-value for each covariate}
   \item \strong{out.cor:} {variable selection results for testing if a main covariate has an effect on the response variable, but NOT accounting for the additional fixed covariate z}
   \item \strong{out.parcor:} {variable selection results for testing if a main covariate has an effect on the response variable, but AFTER accounting for the additional fixed covariate z}
   \item \strong{out.benhoch.cor:} {variable selection results from Benjamini-Hochberg adjusted p-values when p-values do not account for the additional fixed covariate z}
   \item \strong{out.benhoch.parcor:} {variable selection results from Benjamini-Hochberg adjusted p-values when p-values account for the additional fixed covariate z}
   \item \strong{out.w:} {variable selection results from weighted lasso}
   \item \strong{alpha:} {level of significance to compare with the q-values}
   \item \strong{alpha.bh:} {level of significance to compare with the Benjamini-Hochberg adjusted p-values}
   \item \strong{delta:} {the multiplier to the number of predictors in the penalized loss function for variable selection. The loss function is defined as: \code{ SSE/(sigma^2) - n + delta*p } where SSE denotes the residual sum of squares, sigma denotes the estimator of the model error variance, n is the sample size and p is the number of predictors in the selected model. }
   \item \strong{cv.delta.w:} {the selected delta from the cross-validation for the weighted lasso with q-values}
   \item \strong{cv.delta.adapt:} {the selected delta from the cross-validation for the adaptive lasso}
   \item \strong{cv.out.w:} {the aggregated result of the cross-validation for the weighted lasso with q-values}
   \item \strong{cv.out.adapt:} {the aggregated result of the cross-validation for the adaptive lasso}
   \item \strong{w.aic.boot:} {variable selection results from AIC and the Cox regression with exclusion frequency-based weights from random partitioning}
   \item \strong{w.bic.boot:} {variable selection results from BIC and the Cox regression with exclusion frequency-based weights from random partitioning}
   \item \strong{w.kmeans.aic.boot:} {variable selection results from AIC and the Cox regression with exclusion frequency-based weights from k-means partitioning}
   \item \strong{w.kmeans.bic.boot:} {variable selection results from BIC and the Cox regression with exclusion frequency-based weights from k-means partitioning}
   \item \strong{w.kquart.aic.boot:} {variable selection results from AIC and the Cox regression with exclusion frequency-based weights from k-quartile partitioning}
   \item \strong{w.kquart.bic.boot:} {variable selection results from BIC and the Cox regression with exclusion frequency-based weights from k-quartile partitioning}
   \item \strong{w.sort.aic.boot:} {variable selection results from AIC and the Cox regression with exclusion frequency-based weights from sorted partitioning}
   \item \strong{w.sort.bic.boot:} {variable selection results from BIC and the Cox regression with exclusion frequency-based weights from sorted partitioning}
}
}
\description{
This package provides functions to perform variable selection with weighted lasso for both linear regression
and the Cox proportional hazards regression. The weights are chosen to direct the variable selection procedure so that covariates that are highly
associated with the response are likely to be selected and covariates that weakly associated with the response
are less likely to be selected.  Association between the response and the covariates is based on
results from simpler linear/Cox regressions between the response and each covariate, and include, for example,
q-values, partial correlation coefficients, and t-statistics of regression coefficients.
}
\examples{
x = matrix(rnorm(100*5, 0, 1),100,5)
z = matrix(rbinom(100, 1, 0.5),100,1)
y = matrix(z[,1] + 2*x[,1] - 2*x[,2] + rnorm(100, 0, 1), 100)
}
\references{
Garcia, T.P. and M¨uller, S. (2016). Cox regression with exclusion frequency-based weights to
identify neuroimaging markers relevant to Huntington’s disease onset. Annals of Applied Statistics, 10, 2130-2156.

Garcia, T.P. and M¨uller, S. (2014). Influence of measures of significance-based weights in the weighted Lasso.
Journal of the Indian Society of Agricultural Statistics (Invited paper), 68, 131-144.

Garcia, T.P., Mueller, S., Carroll, R.J., Dunn, T.N., Thomas, A.P., Adams, S.H., Pillai, S.D., and Walzem, R.L.
(2013). Structured variable selection with q-values. Biostatistics, DOI:10.1093/biostatistics/kxt012.
}
